+++ stac_fastapi/extensions/stac_fastapi/extensions/core/transaction.py
@@ -1,9 +1,9 @@
 """Transaction extension."""
 
-from typing import List, Optional, Type, Union
+from typing import List, Literal, Optional, Type, Union
 
 import attr
-from fastapi import APIRouter, Body, FastAPI
+from fastapi import APIRouter, Body, FastAPI, Header
 from stac_pydantic import Collection, Item, ItemCollection
 from stac_pydantic.shared import MimeTypes
 from starlette.responses import JSONResponse, Response
@@ -14,6 +14,7 @@ from stac_fastapi.api.routes import create_async_endpoint
 from stac_fastapi.types.config import ApiSettings
 from stac_fastapi.types.core import AsyncBaseTransactionsClient, BaseTransactionsClient
 from stac_fastapi.types.extension import ApiExtension
+from stac_fastapi.types.stac import PartialCollection, PartialItem, PatchOperation
 
 
 @attr.s
@@ -31,10 +32,50 @@ class PutItem(ItemUri):
 
 
 @attr.s
+class PatchItem(ItemUri):
+    """Patch Item."""
+
+    patch: Annotated[
+        Union[PartialItem, List[PatchOperation]],
+        Body(),
+    ] = attr.ib(default=None)
+    content_type: Annotated[
+        Optional[
+            Literal[
+                "application/json-patch+json",
+                "application/merge-patch+json",
+                "application/json",
+            ]
+        ],
+        Header(),
+    ] = attr.ib(default="application/json")
+
+
+@attr.s
 class PutCollection(CollectionUri):
     """Update Collection."""
 
     collection: Annotated[Collection, Body()] = attr.ib(default=None)
+
+
+@attr.s
+class PatchCollection(CollectionUri):
+    """Patch Collection."""
+
+    patch: Annotated[
+        Union[PartialCollection, List[PatchOperation]],
+        Body(),
+    ] = attr.ib(default=None)
+    content_type: Annotated[
+        Optional[
+            Literal[
+                "application/json-patch+json",
+                "application/merge-patch+json",
+                "application/json",
+            ]
+        ],
+        Header(),
+    ] = attr.ib(default="application/json")
 
 
 @attr.s
@@ -114,6 +155,31 @@ class TransactionExtension(ApiExtension):
             endpoint=create_async_endpoint(self.client.update_item, PutItem),
         )
 
+    def register_patch_item(self):
+        """Register patch item endpoint (PATCH
+        /collections/{collection_id}/items/{item_id})."""
+        self.router.add_api_route(
+            name="Patch Item",
+            path="/collections/{collection_id}/items/{item_id}",
+            response_model=Item if self.settings.enable_response_models else None,
+            responses={
+                200: {
+                    "content": {
+                        MimeTypes.geojson.value: {},
+                    },
+                    "model": Item,
+                }
+            },
+            response_class=self.response_class,
+            response_model_exclude_unset=True,
+            response_model_exclude_none=True,
+            methods=["PATCH"],
+            endpoint=create_async_endpoint(
+                self.client.patch_item,
+                PatchItem,
+            ),
+        )
+
     def register_delete_item(self):
         """Register delete item endpoint (DELETE
         /collections/{collection_id}/items/{item_id})."""
@@ -135,11 +201,6 @@ class TransactionExtension(ApiExtension):
             methods=["DELETE"],
             endpoint=create_async_endpoint(self.client.delete_item, ItemUri),
         )
-
-    def register_patch_item(self):
-        """Register patch item endpoint (PATCH
-        /collections/{collection_id}/items/{item_id})."""
-        raise NotImplementedError
 
     def register_create_collection(self):
         """Register create collection endpoint (POST /collections)."""
@@ -184,6 +245,30 @@ class TransactionExtension(ApiExtension):
             endpoint=create_async_endpoint(self.client.update_collection, PutCollection),
         )
 
+    def register_patch_collection(self):
+        """Register patch collection endpoint (PATCH /collections/{collection_id})."""
+        self.router.add_api_route(
+            name="Patch Collection",
+            path="/collections/{collection_id}",
+            response_model=Collection if self.settings.enable_response_models else None,
+            responses={
+                200: {
+                    "content": {
+                        MimeTypes.geojson.value: {},
+                    },
+                    "model": Collection,
+                }
+            },
+            response_class=self.response_class,
+            response_model_exclude_unset=True,
+            response_model_exclude_none=True,
+            methods=["PATCH"],
+            endpoint=create_async_endpoint(
+                self.client.patch_collection,
+                PatchCollection,
+            ),
+        )
+
     def register_delete_collection(self):
         """Register delete collection endpoint (DELETE /collections/{collection_id})."""
         self.router.add_api_route(
@@ -205,10 +290,6 @@ class TransactionExtension(ApiExtension):
             endpoint=create_async_endpoint(self.client.delete_collection, CollectionUri),
         )
 
-    def register_patch_collection(self):
-        """Register patch collection endpoint (PATCH /collections/{collection_id})."""
-        raise NotImplementedError
-
     def register(self, app: FastAPI) -> None:
         """Register the extension with a FastAPI application.
 
@@ -221,8 +302,10 @@ class TransactionExtension(ApiExtension):
         self.router.prefix = app.state.router_prefix
         self.register_create_item()
         self.register_update_item()
+        self.register_patch_item()
         self.register_delete_item()
         self.register_create_collection()
         self.register_update_collection()
+        self.register_patch_collection()
         self.register_delete_collection()
         app.include_router(self.router, tags=["Transaction Extension"])